// src/models/enums.ts
export enum EHashAlgorithms {
  SHA256 = 'sha256',
  SHA512 = 'sha512',
}

export enum ELanguages {
  de = 'de',
  en = 'en',
}

export enum EStages {
  local = 'local',
  dev = 'dev',
  demo = 'demo',
  prod = 'prod',
}

export enum EVerificationState {
  VERIFICATION_PENDING = 'VERIFICATION_PENDING',
  VERIFICATION_FAILED = 'VERIFICATION_FAILED',
  VERIFICATION_PARTLY_SUCCESSFUL = 'VERIFICATION_PARTLY_SUCCESSFUL',
  VERIFICATION_SUCCESSFUL = 'VERIFICATION_SUCCESSFUL',
}

export enum EUppStates {
  created = 'created',
  anchored = 'anchored',
}

export enum EInfo {
  START_VERIFICATION_CALL = 'START_VERIFICATION_CALL',
  PROCESSING_VERIFICATION_CALL = 'PROCESSING_VERIFICATION_CALL',
  START_CHECKING_RESPONSE = 'START_CHECKING_RESPONSE',
  RESPONSE_JSON_PARSED_SUCCESSFUL = 'RESPONSE_JSON_PARSED_SUCCESSFUL',
  UPP_HAS_BEEN_FOUND = 'UPP_HAS_BEEN_FOUND',
  NO_BLXTX_FOUND = 'NO_BLXTX_FOUND',
  BLXTXS_FOUND_SUCCESS = 'BLXTXS_FOUND_SUCCESS',
  WARNING_EMPTY_BLXTX_FOUND = 'WARNING_EMPTY_BLXTX_FOUND',
}

export enum EError {
  NO_ERROR = 'NO_ERROR',
  CERTIFICATE_DATA_MISSING = 'CERTIFICATE_DATA_MISSING',
  CERTIFICATE_ID_CANNOT_BE_FOUND = 'CERTIFICATE_ID_CANNOT_BE_FOUND',
  VERIFICATION_FAILED_EMPTY_RESPONSE = 'VERIFICATION_FAILED_EMPTY_RESPONSE',
  VERIFICATION_FAILED_MISSING_SEAL_IN_RESPONSE = 'VERIFICATION_FAILED_MISSING_SEAL_IN_RESPONSE',
  VERIFICATION_FAILED = 'VERIFICATION_FAILED',
  VERIFICATION_CURRENTLY_UNAVAILABLE = 'VERIFICATION_CURRENTLY_UNAVAILABLE',
  VERIFICATION_UNAVAILABLE = 'VERIFICATION_UNAVAILABLE',
  URL_PARAMS_CORRUPT = 'URL_PARAMS_CORRUPT',
  LOCATION_MALFORMED = 'LOCATION_MALFORMED',
  MANDATORY_FIELD_MISSING = 'MANDATORY_FIELD_MISSING',
  FILLING_FORM_WITH_PARAMS_FAILED = 'FILLING_FORM_WITH_PARAMS_FAILED',
  JSON_PARSE_FAILED = 'JSON_PARSE_FAILED',
  JSON_MALFORMED = 'JSON_MALFORMED',
  CANNOT_ACCESS_FORM_FIELD = 'CANNOT_ACCESS_FORM_FIELD',
  MISSING_PROPERTY_IN_UBRICH_VERIFICATION_INSTANCIATION = 'MISSING_PROPERTY_IN_UBRICH_VERIFICATION_INSTANCIATION',
  ELEMENT_FOR_WIDGET_SELECTOR_NOT_FOUND = 'ELEMENT_FOR_WIDGET_SELECTOR_NOT_FOUND',
  MISSING_PARAM_IDS = 'MISSING_PARAM_IDS',
  PARAM_ID_MAPPING_MISSMATCH = 'PARAM_ID_MAPPING_MISSMATCH',
  MISSING_ACCESS_TOKEN = 'MISSING_ACCESS_TOKEN',
  CERTIFICATE_ANCHORED_BY_NOT_AUTHORIZED_DEVICE = 'CERTIFICATE_ANCHORED_BY_NOT_AUTHORIZED_DEVICE',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  UNKNOWN_ERROR = 'UNKNOWN_ERROR',
}